+++ lib/src/adapty_version.dart
@@ -1,1 +1,1 @@
-const String adaptySDKVersion = '4.0.2';
+const String adaptySDKVersion = '4.0.3';